Corrected Prop 1.1
iiii) was the definition of ii), there is nothing to prove.
The second part of the proof doesn't make sense in my opinion so I
replaced it with something similar we did last year. Why is it weird ?
It doesn't follow from the definition of non degeneration that for all
non zero vector x, x^T A is not zero. That's exactly what we need to
prove.