Update chapter00-valeurs-propres.tex
This is not a correction in the course but more of a suggestion. As a student, I found it really confusing that the algebraic multiplicity of lambda_i was represented as g_i (although the choice of the letter g makes it look like it refers to the geometric multiplicity). That's why I suggest to choose g_i for the geometric multiplicity and a_i for the algebraic mutliplicity.